Remove bzip2's shared library
We do this to force static linking to bzip2. This is important in cases
where conda updates the bzip2 package. We don't want to run into an
issue with having the bzip2 shared library removed before unpacking the
new bzip2 package as this would likely cause a crash and a broken
environment.