Rebuild with toolchain 2.2.2
There was an issue with the `toolchain`; where, on a newer macOS
version, it may not pickup the right SDK. For example, the macOS 10.9
SDK was not being picked up on CircleCI, which runs on a newer version
of macOS. Turns out that setting `MACOSX_DEPLOYMENT_TARGET` is not
sufficient in these cases and one needs to specify the macOS SDK
location as well.

As of `toolchain` 2.2.2, the lazy symbol binding issue has been fixed
thanks to the addition of `-isysroot` compiler flag, which specifies the
location of the macos SDK we intend to use. Combined with the facts that
we specify what minimum macOS version we target and we provide the macOS
SDK, this is enough for the build to use that macOS SDK and not pick up
these symbols, which do not appear until later macOS versions.