3.8 update, adds CondaEcosystemModifyDllSearchPath patch
Without this, Python 3.8.1 on Windows is pretty badly broken (it
could not import sqlite for example, or any other modules that
needed DLLs found in Library/bin).

We could have looked into fixing the new patch (which is really just
a cut-down, less functional version of an old patch), but that comes
with various other disadvantages due to missing (and now documented)
things that the CondaEcosystemModifyDllSearchPath patch does; mainly
the `CONDA_DLL_SEARCH_MODIFICATION_` family of env vars. Without that
feature, users with 3rd party buggy software like Amplitube would
suffer import failure when importing an MKL-based numpy (there are many
other instances of this exact class of problem too). The original patch
also caches its inputs to avoid calling the Win32 kernel (an expensive
context switch) when nothing has changed.

This does change the parameters to LoadLibraryEx when a module is
imported, whereas the newer patch does not, still, we simply must change
those parameters if we want to have control of our DLL search paths and
should lobby for upstream to revert this back to how it was. I expect it
breaks countless 3rd party extension modules that have been relying on
PATH lookup to work for DLLs for the last 20 years.

Also adds a load of build-time tests around this breakage so we catch
this problem if it happens again in the future.

Note: I suspect there is more wrong with Python 3.8 on Windows to do
with threading and startup. I cannot get LIEF to import correctly.