Check CONFIGS env variable in run_osx_build.sh
Without the CONFIGS env variable being set, one will run into an error
such as:

    Invalid value for 'CONFIG_FILE': File './.ci_support/.yaml' does not exist.

This check is in a similar vein as .scripts/run_docker_build.sh.

Also, negative lengths are not allowed in bash 3.x (added in 4.2);
a few lines for string manipiulation are modified for better
bash compatibility.